WIP: Refactor result manager to output_writer and result reader. Implement output_writer. Add path_gen.py with some of the utility functions for managing the results.
