+++ tests/test_inference.py
@@ -5,7 +5,7 @@ import numpy as np
 from eulerpi.core.dense_grid import generate_regular_grid
 from eulerpi.core.inference import InferenceType, inference
 from eulerpi.core.kde import calc_kernel_width, eval_kde_gauss
-from eulerpi.core.result_manager import ResultManager
+from eulerpi.core.result_managers.output_writer import ResultManager
 from eulerpi.examples.simple_models import LinearODE
